Implement a priority based ProductionQueueManager for the AI
The production queue manager is to be used to handle all production
orders from the AI. It keeps track of priorities assigned to items in
the queue and orders the production queue accordingly.

For the code to work, there may be no external calls to
-fo.issueEnqueueBuildingProductionOrder
-fo.issueEnqueueShipProductionOrder
-fo.issueRequeueProductionOrder
-fo.issueDequeueProductionOrder
or the state of the queue becomes corrupted.

At the beginning of each turn,
ProductionQueueManager.update_for_new_turn() must be called to update
the internal state to the ingame production queue and process the
progress made since last turn.

Only a single instance should be created and stored in AIstate.
Appropriate __setstate__ and __getstate__ methods for savegame-pickling
are provided by the class.