Implement a CombinedShipStats class for improved fleet rating calculations.
The class is used to simulate a virtual ship with combined metrics of multiple ships. It can be used to e.g. calculate the rating of a fleet in which case it should yield similar (but not identical) results to combine_ratings() of the the individual ship_ratings.

The benefits of this class are as follows:
1) It offers an alternative approach to calculate the combined strength of multiple ships which may be more accurate to the previous assessment, when the assumptions of combine_ratings (quadratic approximation; similar damage and structure) are not satisfied.
2) It allows us to effectively rate a fleet against another fleet: First calculate the CombinedShipStats for each fleet, then calculate rating with enemy_stats set accordingly.

Future improvements should consider
- number of ships vs number of fighters to estimate number of fighter hits
- better estimation of the effective shield value of a fleet